eta moet geassocieerd worden met long en niet met double
git-svn-id: file:///home/kommusoft/tmpgit/swopsvn@370 c49496ef-df08-40b3-bdd7-5dbf4c3ba35c

+++ ProjectSWOP20102011/src/projectswop20102011/userinterface/DispatchUnitsUserInterface.java
@@ -58,7 +58,7 @@ public class DispatchUnitsUserInterface extends CommandUserInterface {
 					for (int i = 0; i < suggestedUnits.size(); i++) {
 						Unit u = suggestedUnits.get(i);
 						double distance = u.getCurrentLocation().getDistanceTo(selectedEmergency.getLocation());
-						double eta = u.getETA(selectedEmergency.getLocation());
+						long eta = u.getETA(selectedEmergency.getLocation());
 						this.writeOutput(String.format("\t%s\t%s\t%s\t%s\t%s", i, u.getClass().getSimpleName(), u.getName(), distance, eta));
 					}
 					boolean acceptSuggestion = this.parseInputToType(new BooleanParser(), "Do you accept the suggestion");
